Create Basic_SQLi_XSS_detection.py
Modify as needed. More so intended for CTF but has the capability to be modified for a larger dictionary of vulnerabilities for bug bounty use. 

Goal: This program defines two functions for scanning for XSS and time-based SQLi vulnerabilities, and executes the main code block that calls these functions and prints output messages to the console.